Moved section files into their own folder
