+++ frontend/ohio-house-scraper-app/src/components/sections/Footer.js
